// src/protos/ciot/proto/v2/http_client.ts
export interface HttpClientCfg {
    /**
     * @generated from protobuf field: string url = 1
     */
    url: string; // URL for the HTTP request.
    /**
     * @generated from protobuf field: Ciot.HttpClientMethod method = 2
     */
    method: HttpClientMethod; // HTTP method to use.
    /**
     * @generated from protobuf field: Ciot.HttpClientTransportType transport = 3
     */
    transport: HttpClientTransportType; // Transport type for the HTTP request.
    /**
     * @generated from protobuf field: uint32 timeout = 4
     */
    timeout: number; // Timeout for the HTTP request.
}
/**
 * Message representing HTTP client status.
 *
 * @generated from protobuf message Ciot.HttpClientStatus
 */
export interface HttpClientStatus {
    /**
     * @generated from protobuf field: Ciot.HttpClientState state = 1
     */
    state: HttpClientState; // State of the HTTP client.
    /**
     * @generated from protobuf field: uint32 error = 2
     */
    error: number; // Error code, if any.
}
// ...
export enum HttpClientState {
    /**
     * The HTTP client is idle.
     *
     * @generated from protobuf enum value: HTTP_CLIENT_STATE_IDLE = 0;
     */
    IDLE = 0,
    /**
     * The HTTP client has started.
     *
     * @generated from protobuf enum value: HTTP_CLIENT_STATE_STARTED = 1;
     */
    STARTED = 1,
    /**
     * The HTTP client is connecting.
     *
     * @generated from protobuf enum value: HTTP_CLIENT_STATE_CONNECTING = 2;
     */
    CONNECTING = 2,
    /**
     * The HTTP client is connected.
     *
     * @generated from protobuf enum value: HTTP_CLIENT_STATE_CONNECTED = 3;
     */
    CONNECTED = 3,
    /**
     * Data has been received by the HTTP client.
     *
     * @generated from protobuf enum value: HTTP_CLIENT_STATE_DATA_RECEIVED = 4;
     */
    DATA_RECEIVED = 4,
    /**
     * The HTTP client has timed out.
     *
     * @generated from protobuf enum value: HTTP_CLIENT_STATE_TIMEOUT = 5;
     */
    TIMEOUT = 5,
    /**
     * The HTTP client encountered an error.
     *
     * @generated from protobuf enum value: HTTP_CLIENT_STATE_ERROR = 6;
     */
    ERROR = 6
}
/**
 * Enum representing different HTTP methods.
 *
 * @generated from protobuf enum Ciot.HttpClientMethod
 */
export enum HttpClientMethod {
    /**
     * HTTP GET method.
     *
     * @generated from protobuf enum value: HTTP_METHOD_GET = 0;
     */
    HTTP_METHOD_GET = 0,
    /**
     * HTTP POST method.
     *
     * @generated from protobuf enum value: HTTP_METHOD_POST = 1;
     */
    HTTP_METHOD_POST = 1,
    /**
     * HTTP PUT method.
     *
     * @generated from protobuf enum value: HTTP_METHOD_PUT = 2;
     */
    HTTP_METHOD_PUT = 2,
    /**
     * HTTP PATCH method.
     *
     * @generated from protobuf enum value: HTTP_METHOD_PATCH = 3;
     */
    HTTP_METHOD_PATCH = 3,
    /**
     * HTTP DELETE method.
     *
     * @generated from protobuf enum value: HTTP_METHOD_DELETE = 4;
     */
    HTTP_METHOD_DELETE = 4,
    /**
     * HTTP HEAD method.
     *
     * @generated from protobuf enum value: HTTP_METHOD_HEAD = 5;
     */
    HTTP_METHOD_HEAD = 5,
    /**
     * HTTP NOTIFY method.
     *
     * @generated from protobuf enum value: HTTP_METHOD_NOTIFY = 6;
     */
    HTTP_METHOD_NOTIFY = 6,
    /**
     * HTTP SUBSCRIBE method.
     *
     * @generated from protobuf enum value: HTTP_METHOD_SUBSCRIBE = 7;
     */
    HTTP_METHOD_SUBSCRIBE = 7,
    /**
     * HTTP UNSUBSCRIBE method.
     *
     * @generated from protobuf enum value: HTTP_METHOD_UNSUBSCRIBE = 8;
     */
    HTTP_METHOD_UNSUBSCRIBE = 8,
    /**
     * HTTP OPTIONS method.
     *
     * @generated from protobuf enum value: HTTP_METHOD_OPTIONS = 9;
     */
    HTTP_METHOD_OPTIONS = 9,
    /**
     * HTTP COPY method.
     *
     * @generated from protobuf enum value: HTTP_METHOD_COPY = 10;
     */
    HTTP_METHOD_COPY = 10,
    /**
     * HTTP MOVE method.
     *
     * @generated from protobuf enum value: HTTP_METHOD_MOVE = 11;
     */
    HTTP_METHOD_MOVE = 11,
    /**
     * HTTP LOCK method.
     *
     * @generated from protobuf enum value: HTTP_METHOD_LOCK = 12;
     */
    HTTP_METHOD_LOCK = 12,
    /**
     * HTTP UNLOCK method.
     *
     * @generated from protobuf enum value: HTTP_METHOD_UNLOCK = 13;
     */
    HTTP_METHOD_UNLOCK = 13,
    /**
     * HTTP PROPFIND method.
     *
     * @generated from protobuf enum value: HTTP_METHOD_PROPFIND = 14;
     */
    HTTP_METHOD_PROPFIND = 14,
    /**
     * HTTP PROPPATCH method.
     *
     * @generated from protobuf enum value: HTTP_METHOD_PROPPATCH = 15;
     */
    HTTP_METHOD_PROPPATCH = 15,
    /**
     * HTTP MKCOL method.
     *
     * @generated from protobuf enum value: HTTP_METHOD_MKCOL = 16;
     */
    HTTP_METHOD_MKCOL = 16,
    /**
     * Maximum value for HTTP methods.
     *
     * @generated from protobuf enum value: HTTP_METHOD_MAX = 17;
     */
    HTTP_METHOD_MAX = 17
}
/**
 * Enum representing different HTTP transport types.
 *
 * @generated from protobuf enum Ciot.HttpClientTransportType
 */
export enum HttpClientTransportType {
    /**
     * Unknown transport type.
     *
     * @generated from protobuf enum value: HTTP_TRANSPORT_UNKNOWN = 0;
     */
    HTTP_TRANSPORT_UNKNOWN = 0,
    /**
     * TCP transport.
     *
     * @generated from protobuf enum value: HTTP_TRANSPORT_TCP = 1;
     */
    HTTP_TRANSPORT_TCP = 1,
    /**
     * SSL transport.
     *
     * @generated from protobuf enum value: HTTP_TRANSPORT_SSL = 2;
     */
    HTTP_TRANSPORT_SSL = 2
}
// ...
class HttpClientCfg$Type extends MessageType<HttpClientCfg> {
    constructor() {
        super("Ciot.HttpClientCfg", [
            { no: 1, name: "url", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "method", kind: "enum", T: () => ["Ciot.HttpClientMethod", HttpClientMethod] },
            { no: 3, name: "transport", kind: "enum", T: () => ["Ciot.HttpClientTransportType", HttpClientTransportType] },
            { no: 4, name: "timeout", kind: "scalar", T: 13 /*ScalarType.UINT32*/ }
        ]);
    }
    create(value?: PartialMessage<HttpClientCfg>): HttpClientCfg {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.url = "";
        message.method = 0;
        message.transport = 0;
        message.timeout = 0;
        if (value !== undefined)
            reflectionMergePartial<HttpClientCfg>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: HttpClientCfg): HttpClientCfg {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string url */ 1:
                    message.url = reader.string();
                    break;
                case /* Ciot.HttpClientMethod method */ 2:
                    message.method = reader.int32();
                    break;
                case /* Ciot.HttpClientTransportType transport */ 3:
                    message.transport = reader.int32();
                    break;
                case /* uint32 timeout */ 4:
                    message.timeout = reader.uint32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: HttpClientCfg, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string url = 1; */
        if (message.url !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.url);
        /* Ciot.HttpClientMethod method = 2; */
        if (message.method !== 0)
            writer.tag(2, WireType.Varint).int32(message.method);
        /* Ciot.HttpClientTransportType transport = 3; */
        if (message.transport !== 0)
            writer.tag(3, WireType.Varint).int32(message.transport);
        /* uint32 timeout = 4; */
        if (message.timeout !== 0)
            writer.tag(4, WireType.Varint).uint32(message.timeout);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message Ciot.HttpClientCfg
 */
export const HttpClientCfg = new HttpClientCfg$Type();
// @generated message type with reflection information, may provide speed optimized methods
class HttpClientStatus$Type extends MessageType<HttpClientStatus> {
    constructor() {
        super("Ciot.HttpClientStatus", [
            { no: 1, name: "state", kind: "enum", T: () => ["Ciot.HttpClientState", HttpClientState, "HTTP_CLIENT_STATE_"] },
            { no: 2, name: "error", kind: "scalar", T: 13 /*ScalarType.UINT32*/ }
        ]);
    }
    create(value?: PartialMessage<HttpClientStatus>): HttpClientStatus {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.state = 0;
        message.error = 0;
        if (value !== undefined)
            reflectionMergePartial<HttpClientStatus>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: HttpClientStatus): HttpClientStatus {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* Ciot.HttpClientState state */ 1:
                    message.state = reader.int32();
                    break;
                case /* uint32 error */ 2:
                    message.error = reader.uint32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: HttpClientStatus, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* Ciot.HttpClientState state = 1; */
        if (message.state !== 0)
            writer.tag(1, WireType.Varint).int32(message.state);
        /* uint32 error = 2; */
        if (message.error !== 0)
            writer.tag(2, WireType.Varint).uint32(message.error);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message Ciot.HttpClientStatus
 */
export const HttpClientStatus = new HttpClientStatus$Type();